+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.1] - 2019-02-20
+### Changed
+- fixed issue of removing previously added URL rewrites of product when same URL key exists;
+- modified progress bar
+
 ## [1.4.0] - 2019-02-11
 ### Added
 - new option "--entity-type"